requirements.txt
pandas==1.3.5
numpy==1.20.3
sqlalchemy==1.4.22
scipy==1.7.1
backtrader==1.9.76.123
apscheduler==3.9.1
eikon==1.1.15
pandas_market_calendars==3.4
mysql-connector==2.2.9
pymongo==3.12.0